refactor: restructure page types
