+++ packages/types/src/page/gallery.ts
@@ -1,0 +1,1 @@
+export { GalleryComponent, galleryComponentSchema } from "./base";
