+++ packages/types/src/page/hero-banner.ts
@@ -1,0 +1,1 @@
+export { HeroBannerComponent, heroBannerComponentSchema } from "./base";
