+++ packages/types/src/page/index.ts
@@ -1,0 +1,5 @@
+export * from "./base";
+export * from "./hero-banner";
+export * from "./gallery";
+export * from "./value-props";
+export * from "./product-grid";
