+++ packages/types/src/page/product-grid.ts
@@ -1,0 +1,1 @@
+export { ProductGridComponent, productGridComponentSchema } from "./base";
