+++ packages/types/src/page/value-props.ts
@@ -1,0 +1,1 @@
+export { ValuePropsComponent, valuePropsComponentSchema } from "./base";
